# contracts/openapi.json
{
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Structured context for the error, or null.",
            "title": "Detail"
          },
          "error_code": {
            "description": "Machine-readable code. Branch on this, not on the message.",
            "enum": [
              "EMPTY_STATEMENT",
              "MULTIPLE_STATEMENTS",
              "NON_SELECT_STATEMENT",
              "FORBIDDEN_KEYWORD",
              "FORBIDDEN_FUNCTION",
              "INVALID_SQL",
              "QUERY_EXECUTION_ERROR",
              "INVALID_CHART_CONFIG",
              "INVALID_CONNECTION_CONFIG",
              "DB_TLS_REQUIRED",
              "CONNECTION_PAUSED",
              "ROW_LIMIT_EXCEEDED",
              "SQL_TOO_LONG",
              "RESULT_TOO_LARGE",
              "WEAK_PASSWORD",
              "DB_AUTH_FAILED",
              "DB_PERMISSION_DENIED",
              "INVALID_CREDENTIALS",
              "NOT_AUTHENTICATED",
              "ACCOUNT_LOCKED",
              "PASSWORD_CHANGE_REQUIRED",
              "FORBIDDEN",
              "CONNECTION_NOT_FOUND",
              "QUERY_NOT_FOUND",
              "TABLE_NOT_FOUND",
              "DASHBOARD_NOT_FOUND",
              "USER_NOT_FOUND",
              "DUPLICATE_NAME",
              "DUPLICATE_EMAIL",
              "LAST_ADMIN",
              "QUERY_FROZEN",
              "REQUEST_VALIDATION_ERROR",
              "RATE_LIMITED",
              "DB_UNREACHABLE",
              "QUERY_TIMEOUT",
              "INTERNAL_ERROR",
              "SERVICE_NOT_READY"
            ],
            "title": "Error Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "error_code",
          "message"
        ],
        "title": "ApiError",
        "type": "object"
      },
...
      "ColumnInfo": {
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "nullable": {
            "title": "Nullable",
            "type": "boolean"
          },
          "primary_key": {
            "default": false,
            "title": "Primary Key",
            "type": "boolean"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "type",
          "nullable"
        ],
        "title": "ColumnInfo",
        "type": "object"
      },
      "ColumnList": {
        "properties": {
          "columns": {
            "items": {
              "$ref": "#/components/schemas/ColumnInfo"
            },
            "title": "Columns",
            "type": "array"
          },
          "connection_id": {
            "title": "Connection Id",
            "type": "string"
          },
          "table": {
            "title": "Table",
            "type": "string"
          }
        },
        "required": [
          "connection_id",
          "table",
          "columns"
        ],
        "title": "ColumnList",
        "type": "object"
      },
      "ConnectionCreate": {
        "properties": {
          "database": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Database"
          },
          "db_type": {
            "$ref": "#/components/schemas/DbType"
          },
          "host": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Host"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "password": {
            "anyOf": [
              {
                "format": "password",
                "type": "string",
                "writeOnly": true
              },
              {
                "type": "null"
              }
            ],
            "title": "Password"
          },
          "port": {
            "anyOf": [
              {
                "maximum": 65535.0,
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Port"
          },
          "sqlite_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sqlite Path"
          },
          "ssl_mode": {
            "$ref": "#/components/schemas/SslMode",
            "default": "require"
          },
          "ssl_root_cert": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ssl Root Cert"
          },
          "username": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          }
        },
        "required": [
          "name",
          "db_type"
        ],
        "title": "ConnectionCreate",
        "type": "object"
      },
      "ConnectionCreateResult": {
        "description": "A create always returns the saved profile, successful test or not.",
        "properties": {
          "connection": {
            "$ref": "#/components/schemas/ConnectionRead"
          },
          "test_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Test Error"
          },
          "test_error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Test Error Code"
          },
          "test_ok": {
            "title": "Test Ok",
            "type": "boolean"
          }
        },
        "required": [
          "connection",
          "test_ok"
        ],
        "title": "ConnectionCreateResult",
        "type": "object"
      },
...
      "ConnectionRead": {
        "description": "Safe projection of a connection. Has no field capable of holding a secret.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "database": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Database"
          },
          "db_type": {
            "$ref": "#/components/schemas/DbType"
          },
          "host": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Host"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_test_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Test Error"
          },
          "last_tested_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Tested At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "paused": {
            "default": false,
            "title": "Paused",
            "type": "boolean"
          },
          "port": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Port"
          },
          "sqlite_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sqlite Path"
          },
          "ssl_mode": {
            "$ref": "#/components/schemas/SslMode"
          },
          "ssl_root_cert": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ssl Root Cert"
          },
          "status": {
            "$ref": "#/components/schemas/ConnectionStatus"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "username": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          }
        },
        "required": [
          "id",
          "name",
          "db_type",
          "host",
          "port",
          "database",
          "username",
          "sqlite_path",
          "ssl_mode",
          "ssl_root_cert",
          "status",
          "last_tested_at",
          "last_test_error",
          "created_at",
          "updated_at"
        ],
        "title": "ConnectionRead",
        "type": "object"
      },
      "ConnectionStatus": {
        "enum": [
          "untested",
          "ok",
          "failed"
        ],
        "title": "ConnectionStatus",
        "type": "string"
      },
...
      "DbType": {
        "enum": [
          "postgres",
          "mysql",
          "sqlite"
        ],
        "title": "DbType",
        "type": "string"
      },
...
      "SslMode": {
        "description": "How much TLS a target connection insists on.\n\nlibpq's own vocabulary, so the Postgres mapping is the identity function and\nnobody has to hold a translation table in their head while reading a\nconnection row. MySQL has no equivalent parameter and is mapped in\n``app.db.target_registry.mysql_connect_args``.\n\nOrdered weakest to strongest. The two ``verify`` modes are the only ones\nthat authenticate the server rather than merely encrypting the wire, so they\nare the only ones a root certificate applies to.",
        "enum": [
          "disable",
          "allow",
          "prefer",
          "require",
          "verify-ca",
          "verify-full"
        ],
        "title": "SslMode",
        "type": "string"
      },
      "TableInfo": {
        "properties": {
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "kind"
        ],
        "title": "TableInfo",
        "type": "object"
      },
      "TableList": {
        "properties": {
          "connection_id": {
            "title": "Connection Id",
            "type": "string"
          },
          "tables": {
            "items": {
              "$ref": "#/components/schemas/TableInfo"
            },
            "title": "Tables",
            "type": "array"
          }
        },
        "required": [
          "connection_id",
          "tables"
        ],
        "title": "TableList",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Schema-agnostic backend for saved, read-only SQL exposed as chart-ready JSON. All fraud logic lives in the SQL you save.",
    "title": "Fraud Analyzer Engine",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/connections": {
      "get": {
        "description": "List every connection. Credentials are never included.",
        "operationId": "list_connections_connections_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConnectionRead"
                  },
                  "title": "Response List Connections Connections Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          }
        },
        "summary": "List Connections",
        "tags": [
          "connections"
        ]
      },
      "post": {
        "description": "Create a connection profile and test it immediately.\n\nA failing test still saves the profile, with ``status=\"failed\"`` and the\nerror attached, so credentials can be corrected without re-entering\neverything.",
        "operationId": "create_connection_connections_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionCreateResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          }
        },
        "summary": "Create Connection",
        "tags": [
          "connections"
        ]
      }
    },
...
    "/connections/{connection_id}/tables": {
      "get": {
        "description": "List every table and view visible to this connection's credentials.",
        "operationId": "list_tables_connections__connection_id__tables_get",
        "parameters": [
          {
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "title": "Connection Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TableList"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          }
        },
        "summary": "List Tables",
        "tags": [
          "introspection"
        ]
      }
    },
    "/connections/{connection_id}/tables/{table_name}/columns": {
      "get": {
        "description": "List the columns and types of one table or view.",
        "operationId": "list_columns_connections__connection_id__tables__table_name__columns_get",
        "parameters": [
          {
            "in": "path",
            "name": "connection_id",
            "required": true,
            "schema": {
              "title": "Connection Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "table_name",
            "required": true,
            "schema": {
              "title": "Table Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ColumnList"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          },
          "504": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            },
            "description": "Structured error. Branch on `error_code`."
          }
        },
        "summary": "List Columns",
        "tags": [
          "introspection"
        ]
      }
    },
...
  }
}